add security user and room

+++ src/main/java/org/ecommerce/travelappbackend/repository/DestinationImageRepository.java
@@ -4,5 +4,5 @@ package org.ecommerce.travelappbackend.repository;
 import org.ecommerce.travelappbackend.entity.DestinationImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DestinationImageRepository extends JpaRepository<DestinationImage, Integer>{
+public interface DestinationImageRepository extends JpaRepository<DestinationImage, Long>{
 }
